Add files via upload
Today a new method of integration was add, the Runge kutta method. This was an aim to show that the pendulum conserved energy better than using the ~Euler method of integration. The plot is slightly different for each but total energy oscilates slightly. This will be looke at in the next workshop and this method of integration will be added to the double pendulum